LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-circle-expected.txt


FAIL Test for checking position of the svg element when animateMotion's mpath is a circle assert_approx_equals: expected -200 +/- 1 but got 0
